gs/WSPbot/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.153.0
python-dotenv==1.2.1
